+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 This file documents all notable changes to the GEOS-Chem repository starting in version 14.0.0, including all GEOS-Chem Classic and GCHP run directory updates.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [Unreleased] - TBD
+### Added
+- Added 'checkYamlSettings.sh' script in run/GCHP/utils to catch file setting errors in ExtData-2G-compatible yaml files.
 
 ## [14.7.1] - 2026-04-08
 ### Added